Moving dismissed toast to settings
