+++ frontend/src/metabase-types/api/mocks/user.ts
@@ -20,7 +20,6 @@ export const createMockUser = (opts?: Partial<User>): User => ({
   date_joined: new Date().toISOString(),
   first_login: new Date().toISOString(),
   last_login: new Date().toISOString(),
-  dismissed_toasts: {},
   ...opts,
 });
