+++ frontend/src/metabase-types/api/user.ts
@@ -31,7 +31,6 @@ export interface User extends BaseUser {
   custom_homepage: {
     dashboard_id: DashboardId;
   } | null;
-  dismissed_toasts: Record<string, boolean>;
 }
 
 export interface UserListResult {
