+++ frontend/src/metabase/reducers-common.js
@@ -4,7 +4,7 @@
 import app from "metabase/redux/app";
 import embed from "metabase/redux/embed";
 import requests from "metabase/redux/requests";
-import settings from "metabase/redux/settings";
+import { settings } from "metabase/redux/settings";
 import undo from "metabase/redux/undo";
 import upload from "metabase/redux/uploads";
 import { reducer as auth } from "metabase/redux/auth";
