+++ frontend/src/metabase/redux/settings.ts
@@ -1,19 +1,21 @@
+import { createAsyncThunk } from "@reduxjs/toolkit";
 import MetabaseSettings from "metabase/lib/settings";
-
 import {
   handleActions,
   createThunkAction,
   combineReducers,
 } from "metabase/lib/redux";
 
-import { SessionApi } from "metabase/services";
+import { SessionApi, SettingsApi } from "metabase/services";
+import type { ToastKeys } from "metabase-types/api";
+import { Dispatch, GetState } from "metabase-types/store";
 
 export const REFRESH_SITE_SETTINGS = "metabase/settings/REFRESH_SITE_SETTINGS";
 
 export const refreshSiteSettings = createThunkAction(
   REFRESH_SITE_SETTINGS,
-  ({ locale } = {}) =>
-    async (dispatch, getState) => {
+  ({ locale }: { locale?: string } = {}) =>
+    async (dispatch: Dispatch, getState: GetState) => {
       const settings = await SessionApi.properties(null, {
         headers: locale ? { "X-Metabase-Locale": locale } : {},
       });
@@ -32,6 +34,18 @@ const values = handleActions(
   window.MetabaseBootstrap,
 );
 
-export default combineReducers({
+export const DISMISS_TOAST = "metabase/settings/DISMISS_TOAST";
+export const dismissToast = createAsyncThunk(
+  DISMISS_TOAST,
+  async (toastKey: ToastKeys, { dispatch }) => {
+    await SettingsApi.put({
+      key: toastKey,
+      value: true,
+    });
+    await dispatch(refreshSiteSettings());
+  },
+);
+
+export const settings = combineReducers({
   values,
 });
