+++ frontend/src/metabase/redux/user.js
@@ -1,10 +1,9 @@
-import { assocIn } from "icepick";
 import {
   createAction,
   createThunkAction,
   handleActions,
 } from "metabase/lib/redux";
-import { SettingsApi, UserApi } from "metabase/services";
+import { UserApi } from "metabase/services";
 import { CLOSE_QB_NEWB_MODAL } from "metabase/query_builder/actions";
 import Users from "metabase/entities/users";
 
@@ -30,33 +29,12 @@ export const loadCurrentUser = createThunkAction(
 export const CLEAR_CURRENT_USER = "metabase/user/CLEAR_CURRENT_USER";
 export const clearCurrentUser = createAction(CLEAR_CURRENT_USER);
 
-export const DISMISS_TOAST = "metabase/user/DISMISS_TOAST";
-export const dismissToast = createThunkAction(
-  DISMISS_TOAST,
-  toastKey => async (dispatch, getState) => {
-    await SettingsApi.put({
-      key: toastKey,
-      value: true,
-    });
-    return { toastKey };
-  },
-);
-
 export const currentUser = handleActions(
   {
     [CLEAR_CURRENT_USER]: { next: (state, payload) => null },
     [REFRESH_CURRENT_USER]: { next: (state, { payload }) => payload },
     [CLOSE_QB_NEWB_MODAL]: {
       next: (state, { payload }) => ({ ...state, is_qbnewb: false }),
-    },
-    [DISMISS_TOAST]: {
-      next: (state, { payload }) => {
-        if (payload.toastKey) {
-          return assocIn(state, ["dismissed_toasts", payload.toastKey], true);
-        } else {
-          return state;
-        }
-      },
     },
     [Users.actionTypes.UPDATE]: {
       next: (state, { payload }) => {
