+++ frontend/src/metabase/selectors/user.ts
@@ -30,8 +30,3 @@ export const getUserPersonalCollectionId = createSelector(
   [getUser],
   user => user?.personal_collection_id,
 );
-
-export const hasUserDismissedToast = createSelector(
-  [getUser, (state, toastKey) => toastKey],
-  (user, toastKey) => user?.dismissed_toasts[toastKey],
-);
